+++ source/extensions/lightspeed.trex.properties_pane.logic.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0]
+### Added
+- Added Delete button for logic graphs in properties panel
+
 ## [1.1.0]
 ### Added
 - Added create and edit logic graph buttons to properties panel